changed some queries

+++ P2P/plugin/communication/fermat-p2p-plugin-communication-network-node-bitdubai/src/main/java/com/bitdubai/fermat_p2p_plugin/layer/communications/network/node/developer/bitdubai/version_1/structure/catalog_propagation/actors/PropagateActorsCatalogTask.java
@@ -79,13 +79,13 @@ public class PropagateActorsCatalogTask implements Runnable {
         ActorCatalogDao actorCatalogDao = JPADaoFactory.getActorCatalogDao();
         NodeCatalogDao nodeCatalogDao = JPADaoFactory.getNodeCatalogDao();
 
-        Integer currentNodesInCatalog = nodeCatalogDao.getCountOfNodesToPropagateWith(networkNodePluginRoot.getIdentity().getPublicKey());
+        Long currentNodesInCatalog = nodeCatalogDao.getCountOfNodesToPropagateWith(networkNodePluginRoot.getIdentity().getPublicKey());
 
         LOG.info("Executing actor propagation: currentNodesInCatalog: "+currentNodesInCatalog);
 
         if (currentNodesInCatalog > 0) {
 
-            Integer countOfItemsToShare = actorCatalogDao.getCountOfItemsToShare(currentNodesInCatalog);
+            Long countOfItemsToShare = actorCatalogDao.getCountOfItemsToShare(currentNodesInCatalog);
 
             LOG.info("Executing actor propagation: countOfItemsToShare: "+countOfItemsToShare);
