+++ P2P/plugin/communication/fermat-p2p-plugin-communication-network-node-bitdubai/src/main/java/com/bitdubai/fermat_p2p_plugin/layer/communications/network/node/developer/bitdubai/version_1/structure/database/jpa/daos/NodeCatalogDao.java
@@ -146,7 +146,7 @@ public class NodeCatalogDao extends AbstractBaseDao<NodeCatalog> {
         }
     }
 
-    public final Integer getCountOfNodesToPropagateWith(final String identityPublicKey) throws CantReadRecordDataBaseException {
+    public final Long getCountOfNodesToPropagateWith(final String identityPublicKey) throws CantReadRecordDataBaseException {
 
         LOG.debug("Executing getCountOfNodesToPropagateWith identityPublicKey (" + identityPublicKey + ")");
 
@@ -158,7 +158,7 @@ public class NodeCatalogDao extends AbstractBaseDao<NodeCatalog> {
                     "FROM NodeCatalog a " +
                     "WHERE a.id <> :id";
 
-            TypedQuery<Integer> q = connection.createQuery(sqlQuery, Integer.class);
+            TypedQuery<Long> q = connection.createQuery(sqlQuery, Long.class);
 
             q.setParameter("id", identityPublicKey);
 
@@ -171,7 +171,7 @@ public class NodeCatalogDao extends AbstractBaseDao<NodeCatalog> {
         }
     }
 
-    public final Integer getCountOfItemsToShare(final Integer currentNodesInCatalog) throws CantReadRecordDataBaseException {
+    public final Long getCountOfItemsToShare(final Long currentNodesInCatalog) throws CantReadRecordDataBaseException {
 
         LOG.debug("Executing getCountOfItemsToShare currentNodesInCatalog (" + currentNodesInCatalog + ")");
 
@@ -183,7 +183,7 @@ public class NodeCatalogDao extends AbstractBaseDao<NodeCatalog> {
                     "FROM NodeCatalog a " +
                     "WHERE a.pendingPropagations > 0 AND a.triedToPropagateTimes > :triedToPropagateTimes";
 
-            TypedQuery<Integer> q = connection.createQuery(sqlQuery, Integer.class);
+            TypedQuery<Long> q = connection.createQuery(sqlQuery, Long.class);
 
             q.setParameter("triedToPropagateTimes", currentNodesInCatalog);
 
@@ -196,7 +196,7 @@ public class NodeCatalogDao extends AbstractBaseDao<NodeCatalog> {
         }
     }
 
-    public final List<NodePropagationInformation> listItemsToShare(final Integer currentNodesInCatalog) throws CantReadRecordDataBaseException {
+    public final List<NodePropagationInformation> listItemsToShare(final Long currentNodesInCatalog) throws CantReadRecordDataBaseException {
 
         LOG.debug("Executing NodeCatalogDao.listItemsToShare currentNodesInCatalog (" + currentNodesInCatalog + ")");
